vim: surround: Don't use feedkeys() for non-human
Old code uses feedkeys() as if it is ungetc() for stdio, but it is not
so.  If typeahead buffer is not empty, the assumption is broken.  So
that new code avoids to use feedkeys().

Note that this change introduce a bug.  When key sequence which consists
of 2 or more keys and which does not match to any surrounding object is
inputted, surrounding process doesn't abort so that it causes unexpected
behavior.